+++ commonmark-integration-test/src/test/java/org/commonmark/integration/SpecIntegrationTest.java
@@ -8,7 +8,7 @@ import org.commonmark.ext.gfm.tables.TablesExtension;
 import org.commonmark.ext.front.matter.YamlFrontMatterExtension;
 import org.commonmark.renderer.html.HtmlRenderer;
 import org.commonmark.parser.Parser;
-import org.commonmark.testutil.spec.SpecExample;
+import org.commonmark.testutil.example.Example;
 import org.commonmark.testutil.SpecTestCase;
 import org.junit.Test;
 
@@ -30,7 +30,7 @@ public class SpecIntegrationTest extends SpecTestCase {
     private static final HtmlRenderer RENDERER = HtmlRenderer.builder().extensions(EXTENSIONS).percentEncodeUrls(true).build();
     private static final Map<String, String> OVERRIDDEN_EXAMPLES = getOverriddenExamples();
 
-    public SpecIntegrationTest(SpecExample example) {
+    public SpecIntegrationTest(Example example) {
         super(example);
     }