+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html),
 with the exception that 0.x versions can break between minor versions.
 
+## Unreleased
+### Changed
+- Parse backslash followed by unescapable character the same way as
+  the reference implementations.
+### Fixed
+- Fix tab handling in ATX and Setext headings.
+
 ## [0.11.0] - 2018-01-17
 ### Added
 - The extension for tables now also renders to plain text